Forgot these files. Still the same fixes Changed the interactive script to use kdialog, instead of in the terminal to be more user friendly.
Removed an unneccisary work around dealing with /etc/ld.so.cache. The imported apps work without it, and it was messing with apt-get updates updating the cache, as it renames the file, instead of writing directly to it. It was mounting over /etc/ld.so.cache, (in seperate namespaces). The file was in use because of this, and the update of the cache fails, causing apt-get to also fail.

The shutdown prompt was the only fix applied in revision 6...

git-svn-id: svn://svn.code.sf.net/p/linuxrcd/code@7 3df0df1f-9518-46d5-b0fc-6e62cabb456f