Another backup commit. This should be the last changes. Now it time to make the translation changes to the files themselves
This allows all filenames and file contents containing a specified sting to have that string replaced with a new one

this is in recoverylivecdmaker_files/usr/share/linuxrcd/translations/English/* (excluding the directories) the file name is the string you want replaced, and the contents of the file is what you want to replace the string with. (to ensure that you add no white space at the end you can add 11 tildes like this ~~~~~~~~~~~ , it has to be eleven

the contents of the file in recoverylivecdmaker_files/usr/share/linuxrcd/translations/English/@Language_Name@ could be 
English~~~~~~~~~~~ and every instance of the string @Language_Name@ in the file/folder's names or contents would be replaced with the string English (it will not add in the 11 tildes in the file )


-----------------------------------------------------------
this also allows the contents, and name of one specified file to be replaced as well

This is in /recoverylivecdmaker_files/usr/share/linuxrcd/translations/English/file_translations/*

If I wanted to translate every occurance of @Tool_Name@ in the file /recoverylivecdmaker_files/usr/bin/@greeting@.sh, I would create a file /recoverylivecdmaker_files/usr/share/linuxrcd/translations/English/file_translations/usr/bin/@greeting@.sh/TRANSLATION_DATA notice that after /recoverylivecdmaker_files/usr/share/linuxrcd/translations/English/file_translations, is the path to the file, only it is a folder /usr/bin/@greeting@.sh

I would add a line in the file:
@Tool_Name@~~~~~~~~~~~Grub Fixer
notice it has to be 11 tildes, and is space freindly. the only thing not supported as translation text is multiple lines

If I wanted this file to be renamed I could add a file /recoverylivecdmaker_files/usr/share/linuxrcd/translations/English/FILEFOLDERNAME

with the contents 
Welcome.sh~~~~~~~~~~~

Or I can make /recoverylivecdmaker_files/usr/share/linuxrcd/translations/English/FILEFOLDERNAME have the same name of the file
@greeting@.sh

and add another line in /recoverylivecdmaker_files/usr/share/linuxrcd/translations/English/file_translations/usr/bin/@greeting@.sh/TRANSLATION_DATA

@greeting@~~~~~~~~~~~welcome

and it will replace the string @greeting@ in /recoverylivecdmaker_files/usr/share/linuxrcd/translations/English/FILEFOLDERNAME with welcome, resulting in the string

welcome.sh

and it will rename the file to that.

note you can do the same with folder names. 

ten tildes are a deleminator. you cant have less then eleven, because it will see them as part of one string, and you cant have more because tildes will start appearing in strings. If you have 22, then it appears as empty.

git-svn-id: svn://svn.code.sf.net/p/linuxrcd/code@12 3df0df1f-9518-46d5-b0fc-6e62cabb456f