Backup commit. This commit was delayed because the charger went again for my peice of 2004 laptop that I was writing this on...
Allow translations to have regex chars. (execept for the " char, as it stops kdialog) Will fix by next commit

Added English translations. The whole files translatiabilty may need to be changed, as in names of the translation placeholders and comments. once thats done it will be turned into a template

still using lxterminal...

git-svn-id: svn://svn.code.sf.net/p/linuxrcd/code@14 3df0df1f-9518-46d5-b0fc-6e62cabb456f