Fixed issues with triailing, deliminated space stoping sed from working. If a string ends in a space, it will be clipped.
I think the translatability of the files is done. all that would be left here is polishing up the files to make templates, and a small script to be more user friendly to the translator. 

kdialog gave a few issues with bash completion. so sticking with lxterminal. the issue where after a while, lxterminal does not like enter and tabs is probably due to the hacky workaround I use to get it to run in the users system without perminantly installing it.

title bars of kdialog windows now have custom (translatable) titles



git-svn-id: svn://svn.code.sf.net/p/linuxrcd/code@16 3df0df1f-9518-46d5-b0fc-6e62cabb456f