Backup commit. not fully tested yet, but close. Still, do not use, unless you know what you are doing
Replaced hacky method of using unshare,  with a method of using patchelf from http://hydra.nixos.org/project/patchelf to force the executibles, and its libraries to be searched for in a paticular location using RPATH. And set the interpreter (ld-linux.so) to the one in this folder as well. The files are copied to the location.

This method actually works. All these utilities are ELF executibles, and this method will only work on ELF. Problems might arise if one of the tools are Python or something, but there is probably ways to go around that once needed. All these tools currently are ELF...

/usr/share is union mounted. the data files in /usr/share should not be so version dependant as the libraries and executibles. The imported ones are at the bottom, and the ones in the users systems are writeable, and at the top, so the files on the users system rule out. 

With this there is much more reliablity, and there is probably a wider range now of apps that can be used.
-----
The CPU arch is now specifiable for the entire build process, its a variable set in the launcher file right now, as well as language. these can easily be made into arguments soon...
-----
fixed timeout to be 10 seconds
-----
Xephyr only appears during a recovery. It does not startup with the CD.
-----
Script no longer runs in an xterm, its now hidden since its interface is kdialog.
-----
Added a custom shutdown handler so that the user does not need to worry about using the system default (was actually commited in the last two revisions due to some issues with SVN)


A little more tweaks and testing is needed until a "real" commit, and a new ISO file. It might even be bumped to beta now that the hacky unshare method of importing apps is replaced.

git-svn-id: svn://svn.code.sf.net/p/linuxrcd/code@19 3df0df1f-9518-46d5-b0fc-6e62cabb456f