Begin the in file description for the translations, so that the script to make translations will see them. It makes no sense to put the documentation for the translations in the translations themselves.
Desktop Xephyr caller now tells the user when it fails to load, because the top cause of it is usualy because Xephyr is not open yet.

Set vdso=0 for the kernel boot argument. Some older systems like this when you are chrooting, but Ubuntu 4.10 apps still crash under a normal chroot of of this CD. Even ones like ls.

Import the Pango libraries. These libraries seem to work with different versions.

try to secure the users system from the web browser even more by hiding the mount point in another namespace, and giving the user a strange UID so that there is less of a chance that it has write permissions. 

This required a new user that had the UID of 1000, because casper seems to need one. This new user is called "linuxrcd" and the older user was renamed to "browser"

Changed the file manager. It still needs an xsettings manager.






git-svn-id: svn://svn.code.sf.net/p/linuxrcd/code@21 3df0df1f-9518-46d5-b0fc-6e62cabb456f