Remove wrongly placed translation file
Fix the shutdown situation to allow the user to shut down from the menu even when they have a system open. This is now can happen after I made the shutdown wait on a socket. 

Add a launcher on the recovery desktop so that the user can get the launcher panel back if the close it.





git-svn-id: svn://svn.code.sf.net/p/linuxrcd/code@104 3df0df1f-9518-46d5-b0fc-6e62cabb456f