Fix the translation back to usr. (usr gets changed later on to RCD)
Remove the patchelf hacks.

Find a different way to get the programs "into" the chroot system now that patchelf is no longer needed WITHOUT bringing in the full /bin directory. The full libraries are brought in, but I want to be careful about what programs I bring in, like apt. This should also make the CD smaller as it uses symlinks

git-svn-id: svn://svn.code.sf.net/p/linuxrcd/code@110 3df0df1f-9518-46d5-b0fc-6e62cabb456f